refactor(service): 순환 참조 해결
순환 참조 문제를 해결하기 위해 별도의 클래스, AuthReactivationCommandService를 생성해서 계정 복구 로직을 분리했습니다.
